requirements-dev.txt
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.0
django-bootstrap-form==3.4
djangorestframework==3.9.1
Pillow==5.3.0
python-decouple==3.1
pytz==2018.7
static3==0.7.0